+++ Bugsnag/Payload/BugsnagEvent.m
@@ -188,8 +188,6 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
             return [BugsnagDeviceWithState deviceFromJson:dict];
         }) ?: [[BugsnagDeviceWithState alloc] init];
 
-        _error = BSGDeserializeDict(json[BSGKeyMetadata][BSGKeyError]);
-
         _errors = BSGDeserializeArrayOfObjects(json[BSGKeyExceptions], ^id _Nullable(NSDictionary * _Nonnull dict) {
             return [BugsnagError errorFromJson:dict];
         }) ?: @[];
@@ -285,6 +283,8 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
         metadata = [BugsnagMetadata new];
     }
 
+    [metadata addMetadata:error toSection:BSGKeyError];
+
     // Device information that isn't part of `event.device`
     NSMutableDictionary *deviceMetadata = BSGParseDeviceMetadata(event);
 #if BSG_HAVE_BATTERY
@@ -333,7 +333,7 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
 
     if (errorReportingThread.crashInfoMessage) {
         [errors[0] updateWithCrashInfoMessage:(NSString * _Nonnull)errorReportingThread.crashInfoMessage];
-        error[@"crashInfo"] = errorReportingThread.crashInfoMessage;
+        [metadata addMetadata:errorReportingThread.crashInfoMessage withKey:@"crashInfo" toSection:@"error"];
     }
     
     BugsnagHandledState *handledState;
@@ -393,7 +393,6 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     obj.featureFlagStore = BSGFeatureFlagStoreFromJSON([event valueForKeyPath:@"user.state.client.featureFlags"]);
     obj.context = [event valueForKeyPath:@"user.state.client.context"];
     obj.customException = BSGParseCustomException(event, [errors[0].errorClass copy], [errors[0].errorMessage copy]);
-    obj.error = error;
     obj.depth = depth;
     obj.usage = [event valueForKeyPath:@"user.usage"];
     return obj;
@@ -417,7 +416,6 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
     _context = BSGDeserializeString(json[BSGKeyContext]);
     _featureFlagStore = [[BSGFeatureFlagStore alloc] init];
     _groupingHash = BSGDeserializeString(json[BSGKeyGroupingHash]);
-    _error = [self getMetadataFromSection:BSGKeyError];
 
     if (_errors.count) {
         BugsnagError *error = _errors[0];
@@ -612,8 +610,6 @@ NSDictionary *BSGParseCustomException(NSDictionary *report,
 
     //  Inserted into `context` property
     [metadata removeObjectForKey:BSGKeyContext];
-    // Build metadata
-    metadata[BSGKeyError] = self.error;
 
     // add user
     event[BSGKeyUser] = [self.user toJson];